updated copy component + location pin
