+++ src/styles/text.module.scss
@@ -93,6 +93,7 @@ $lf: $sf + 0.6;
     @include smallFont();
     color: $dark;
     margin: 0;
+    height: 80%;
 }
